refactor: 타입 경로 수정 및 컴포넌트 통합
